fix(frontend): stop breaking the RSC boundary on test run detail
page.tsx (a Server Component) imported fetchSmallTestRunResults from
useTestRunDetailData.ts, which also holds the useTestRunDetailData
hook and its React imports (useCallback/useEffect/useState). Pulling
that whole module into a Server Component's graph broke the build:
"You're importing a module that depends on useEffect into a React
Server Component module."

Split the server-safe fetchers (fetchAllTestResults,
fetchSmallTestRunResults -- no React import) into their own module,
test-run-results.ts, matching the existing comparison-runs.ts pattern
used by the same page for its other prefetch call. page.tsx now
imports from there; useTestRunDetailData.ts keeps only the hook and
gains an explicit 'use client', matching every other hook file in this
directory.
